requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
Django==4.2.19
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
